+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file. This library adheres to a versioning policy described in [the README](./README.md#versioning). The public API of this library consists of the functions exported in [h3core.js](./lib/h3core.js).
 
 ## Unreleased
+### Added
+- Added `reverseDirectedEdge` function
+### Changed
+- Updated the core library to `v4.5.0`
 ### Fixed
 - Prevent oversized `gridDisk` and `gridDiskDistances` calls from corrupting subsequent calls.